requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
blueprint==3.4.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cryptography==41.0.7
Django==4.2.7
django-bootstrap4==23.2
django-crispy-forms==2.1
django-sslserver==0.22
Faker==20.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pymssql==2.2.11
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.9.0
tzdata==2023.3
update==0.0.1
urllib3==2.1.0
Werkzeug==3.0.1
whitenoise==6.6.0
